Revert change that permits using factor variables.
The change was slightly bugged anyways; there are other areas of the code which
use `select_if` which should've also been changed. If this is brought back we
need to use `\(x) any(is.numeric(x), is.factor(x))` in each of those places
which update the explanatoryVariables (and possibly the responseVariable; what
if we are modeling whether XYZ causes A to be true or false?).