Improve the effect of the reset button.
The reset button now causes virtually all application content to revert to the state it was in when a new session is began.

TODO: There is a couple paragraphs of content in the MLR tabPanel which need to be conditioned upon the observable of the calculate button.

Avoiding reactivity like this is really gross.